+++ workers/source-maps/types/source-map-task/index.ts
@@ -1,0 +1,2 @@
+export { SourceMapsEventWorkerTask, SourcemapCollectedData } from './source-maps-event-worker-task';
+export { SourceMapDataExtended, SourceMapFileChunk, SourceMapsRecord } from './source-maps-record';
