+++ workers/source-maps/types/source-map-task/source-maps-record.ts
